Chore: remove unused folders
